Keep CELT encoder state float-width

+++ celt/encoder_test.go
@@ -148,7 +148,7 @@ func TestEncoderMatchesDecoder(t *testing.T) {
 		t.Errorf("PrevEnergy length: enc=%d, dec=%d", len(encEnergy), len(decEnergy))
 	}
 	for i := range encEnergy {
-		if encEnergy[i] != decEnergy[i] {
+		if float64(encEnergy[i]) != decEnergy[i] {
 			t.Errorf("PrevEnergy[%d]: enc=%f, dec=%f", i, encEnergy[i], decEnergy[i])
 		}
 	}
@@ -398,8 +398,8 @@ func TestEncodeFrameLFEClampsHighBandEnergy(t *testing.T) {
 	if len(lastBandLogE) < 3 {
 		t.Fatalf("lastBandLogE length=%d want at least 3", len(lastBandLogE))
 	}
-	baseAbs := lastBandLogE[0] + eMeans[0]*DB6
-	highAbs := lastBandLogE[2] + eMeans[2]*DB6
+	baseAbs := float64(lastBandLogE[0]) + eMeans[0]*DB6
+	highAbs := float64(lastBandLogE[2]) + eMeans[2]*DB6
 	limitAbs := baseAbs + float64(celtLog2(float32(lfeBandClamp)))
 	floorAbs := float64(celtLog2(float32(celtFloatEpsilon)))
 	wantMax := limitAbs
@@ -416,10 +416,10 @@ func TestLastBandLogEUsesGLogWidth(t *testing.T) {
 	enc.lastBandLogE = appendFloat64AsGLog(enc.lastBandLogE, []float64{1.0 / 3.0})
 	enc.lastBandLogE2 = appendFloat64AsGLog(enc.lastBandLogE2, []float64{2.0 / 3.0})
 
-	if got, want := enc.GetLastBandLogE()[0], float64(celtGLog(1.0/3.0)); got != want {
+	if got, want := enc.GetLastBandLogE()[0], celtGLog(1.0/3.0); got != want {
 		t.Fatalf("GetLastBandLogE()[0]=%0.9g want celt_glog %0.9g", got, want)
 	}
-	if got, want := enc.GetLastBandLogE2()[0], float64(celtGLog(2.0/3.0)); got != want {
+	if got, want := enc.GetLastBandLogE2()[0], celtGLog(2.0/3.0); got != want {
 		t.Fatalf("GetLastBandLogE2()[0]=%0.9g want celt_glog %0.9g", got, want)
 	}
 }
